Added data prebuild step.
NYC counties are now taken care of there, and don't have to be
treated specially by the data import process on the web app.

All dates are now included in the data file for all counties,
so that the client never has holes in the records.

A couple of spots, large variables are freed up explicitly now.

Issue: #11, #50